Issue #2256 has been fixed.

+++ packages/Webkul/Admin/src/Http/Controllers/Contact/Persons/PersonController.php
@@ -2,6 +2,7 @@
 
 namespace Webkul\Admin\Http\Controllers\Contact\Persons;
 
+use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Resources\Json\JsonResource;
@@ -138,17 +139,27 @@ class PersonController extends Controller
     {
         $person = $this->personRepository->findOrFail($id);
 
+        if (
+            $person->leads
+            && $person->leads->count() > 0
+        ) {
+            return response()->json([
+                'message' => trans('admin::app.contacts.persons.index.delete-failed'),
+            ], 400);
+        }
+
         try {
-            Event::dispatch('contacts.person.delete.before', $id);
+            Event::dispatch('contacts.person.delete.before', $person);
 
-            $person->delete($id);
+            $person->delete();
 
-            Event::dispatch('contacts.person.delete.after', $id);
+            Event::dispatch('contacts.person.delete.after', $person);
 
             return response()->json([
                 'message' => trans('admin::app.contacts.persons.index.delete-success'),
             ], 200);
-        } catch (\Exception $exception) {
+
+        } catch (Exception $exception) {
             return response()->json([
                 'message' => trans('admin::app.contacts.persons.index.delete-failed'),
             ], 400);
@@ -158,20 +169,43 @@ class PersonController extends Controller
     /**
      * Mass Delete the specified resources.
      */
-    public function massDestroy(MassDestroyRequest $massDestroyRequest): JsonResponse
+    public function massDestroy(MassDestroyRequest $request): JsonResponse
     {
-        $persons = $this->personRepository->findWhereIn('id', $massDestroyRequest->input('indices'));
+        try {
+            $persons = $this->personRepository->findWhereIn('id', $request->input('indices', []));
 
-        foreach ($persons as $person) {
-            Event::dispatch('contact.person.delete.before', $person);
+            $notDeleted = [];
 
-            $this->personRepository->delete($person->id);
+            foreach ($persons as $person) {
+                if (
+                    $person->leads
+                    && $person->leads->count() > 0
+                ) {
+                    $notDeleted[] = $person->name ?? "ID: {$person->id}";
 
-            Event::dispatch('contact.person.delete.after', $person);
+                    continue;
+                }
+
+                Event::dispatch('contact.person.delete.before', $person);
+
+                $this->personRepository->delete($person->id);
+
+                Event::dispatch('contact.person.delete.after', $person);
+            }
+
+            $message = trans('admin::app.contacts.persons.index.delete-success');
+
+            if (! empty($notDeleted)) {
+                $message .= ' '.trans('admin::app.contacts.persons.index.delete-partial-warning', [
+                    'persons' => implode(', ', $notDeleted),
+                ]);
+            }
+
+            return response()->json(['message' => $message]);
+        } catch (Exception $exception) {
+            return response()->json([
+                'message' => trans('admin::app.contacts.persons.index.delete-failed'),
+            ], 400);
         }
-
-        return response()->json([
-            'message' => trans('admin::app.contacts.persons.index.delete-success'),
-        ]);
     }
 }
